Organize folders and filters in VS
