+++ source/DirectXGameEngineSkeleton/Game Engine/constants.h
