+++ source/DirectXGameEngineSkeleton/megaman.h
@@ -2,7 +2,7 @@
 #define _SPACEWAR_H             // ..file is included in more than one place
 #define WIN32_LEAN_AND_MEAN
 
-#include "game.h"
+#include "Game Engine/game.h"
 
 //=============================================================================
 // Create game class
